msm: hotplug: support hotplug with wfi only power collapse
Currently the pen_release is not set to -1 after we return from
msm_pm_cpu_enter_lowpower. This doesn't happen when we actually power
collapse, but it does when we only wfi. This will cause the next hotplug
attempt to fail and crash the system, as the cpu comes up when the
system is not expecting it. Fix this by explicitly setting the
pen_release when we wake up in this manner.

Change-Id: Ib39c406091412b04510df5de660fd6346763f2b8
Signed-off-by: Jeff Ohlstein <johlstei@codeaurora.org>
(cherry picked from commit 766ccf63f869157a6c992bccc4d9f1dc52cb89b8)